Increase size of sent messages to 53 characters
Agathe Battestini, Vidya Setlur and Timothy Sohn state in their paper "A Large Scale Study of Text Messaging Use" that the average length of text messages sent by american students is 47 characters for males and  58 for females. Thus a length of 53 seems to be reasonable for testing and evaluating our results.